refactor: moved overview card to parts folder
